#33 wip move api cache to shared lib
